refactor: common/ui 패키지 common/base 패키지로 네이밍 변경
