+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/base/BaseDialogFragment.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.common.ui
+package us.wedemy.eggeum.android.common.base
 
 import android.content.Context
 import android.graphics.Color
