+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/MainActivity.kt
@@ -18,7 +18,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
-import us.wedemy.eggeum.android.common.ui.BaseActivity
+import us.wedemy.eggeum.android.common.base.BaseActivity
 import us.wedemy.eggeum.android.main.R
 import us.wedemy.eggeum.android.main.databinding.ActivityMainBinding
 import us.wedemy.eggeum.android.main.model.CafeDetailModel
