+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/adapter/CafeImageAdapter.kt
@@ -9,7 +9,7 @@ package us.wedemy.eggeum.android.main.ui.adapter
 
 import android.view.ViewGroup
 import us.wedemy.eggeum.android.common.extension.layoutInflater
-import us.wedemy.eggeum.android.common.ui.BaseRecyclerViewAdapter
+import us.wedemy.eggeum.android.common.base.BaseRecyclerViewAdapter
 import us.wedemy.eggeum.android.main.databinding.ItemCafeImageBinding
 import us.wedemy.eggeum.android.main.model.FileModel
 import us.wedemy.eggeum.android.main.ui.adapter.viewholder.CafeImageViewHolder
