+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/adapter/NewCafeAdapter.kt
@@ -9,8 +9,8 @@ package us.wedemy.eggeum.android.main.ui.adapter
 
 import android.view.ViewGroup
 import us.wedemy.eggeum.android.common.extension.layoutInflater
-import us.wedemy.eggeum.android.common.ui.BaseRecyclerViewAdapter
-import us.wedemy.eggeum.android.common.ui.BaseViewHolder
+import us.wedemy.eggeum.android.common.base.BaseRecyclerViewAdapter
+import us.wedemy.eggeum.android.common.base.BaseViewHolder
 import us.wedemy.eggeum.android.domain.model.place.PlaceEntity
 import us.wedemy.eggeum.android.main.databinding.ItemNewCafeBinding
 import us.wedemy.eggeum.android.main.ui.adapter.listener.SearchCafeClickListener
