+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/adapter/viewholder/NoticeCardViewHolder.kt
@@ -7,7 +7,7 @@
 
 package us.wedemy.eggeum.android.main.ui.adapter.viewholder
 
-import us.wedemy.eggeum.android.common.ui.BaseViewHolder
+import us.wedemy.eggeum.android.common.base.BaseViewHolder
 import us.wedemy.eggeum.android.main.databinding.ItemNoticeCardBinding
 import us.wedemy.eggeum.android.main.model.NoticeCardModel
 import us.wedemy.eggeum.android.main.ui.adapter.listener.NoticeCardClickListener
