+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/myaccount/ReportCompleteFragment.kt
@@ -11,7 +11,7 @@ import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import us.wedemy.eggeum.android.common.extension.safeNavigate
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.main.databinding.FragmentReportCompleteBinding
 
 class ReportCompleteFragment : BaseFragment<FragmentReportCompleteBinding>() {
