+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/myaccount/ReportFragment.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.launch
 import us.wedemy.eggeum.android.common.extension.repeatOnStarted
 import us.wedemy.eggeum.android.common.extension.safeNavigate
 import us.wedemy.eggeum.android.common.extension.textChangesAsFlow
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.main.databinding.FragmentReportBinding
 import us.wedemy.eggeum.android.main.viewmodel.ReportViewModel
