+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/search/MapFragment.kt
@@ -42,7 +42,7 @@ import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.distinctUntilChangedBy
 import kotlinx.coroutines.launch
 import us.wedemy.eggeum.android.common.extension.repeatOnStarted
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.common.util.fadeInView
 import us.wedemy.eggeum.android.common.util.fadeOutView
 import us.wedemy.eggeum.android.domain.model.place.PlaceEntity
