+++ onboard/src/main/kotlin/us/wedemy/eggeum/android/onboard/ui/RegisterAccountFragment.kt
@@ -16,7 +16,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import us.wedemy.eggeum.android.common.extension.repeatOnStarted
 import us.wedemy.eggeum.android.common.extension.safeNavigate
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.onboard.databinding.FragmentRegisterAccountBinding
 import us.wedemy.eggeum.android.onboard.viewmodel.OnBoardViewModel
