+++ update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/ui/SelectOpinionFragment.kt
@@ -16,7 +16,7 @@ import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import us.wedemy.eggeum.android.common.extension.repeatOnStarted
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.updatecafe.databinding.FragmentSelectOpinionBinding
 import us.wedemy.eggeum.android.updatecafe.viewmodel.SelectOpinionViewModel
