+++ update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/ui/UpdateCafeCompleteFragment.kt
@@ -13,7 +13,7 @@ import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.updatecafe.databinding.FragmentUpdateCafeCompleteBinding
 
 @AndroidEntryPoint
