+++ update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/ui/UpdateMenuCompleteFragment.kt
@@ -14,7 +14,7 @@ import android.view.View
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import us.wedemy.eggeum.android.common.extension.safeNavigate
-import us.wedemy.eggeum.android.common.ui.BaseFragment
+import us.wedemy.eggeum.android.common.base.BaseFragment
 import us.wedemy.eggeum.android.updatecafe.databinding.FragmentUpdateMenuCompleteBinding
 
 @AndroidEntryPoint
